Added half-kana to full-kana conversion switch.
Scripts store string literals with (almost) all hiragana converted to
half-width katakana.  (In Shift-JIS, half-kana are one byte, while
hiragana are two.)

The game's mapping is not in any convenient byte order, and the half-width
table is in a different order from full-width in the spec anyway.
htoz_tablegen.py generates the lookup table from the strings used in the
game's decoder.